requirements.txt
# GUI Framework
PyQt6==6.6.1
PyQt6-Qt6==6.6.1

# Audio processing
pygame-ce==2.5.2
pyaudio==0.2.14
mutagen==1.47.0

# Google Cloud Services
google-cloud-speech==2.26.0
google-cloud-aiplatform==1.73.0
google-cloud-storage==2.14.0

# Utilities
python-dotenv==1.0.0
